pull exceptions out into own file
